UploadPhoto components added
