+++ frontend/src/components/Buttons/CLButtons.jsx
@@ -1,8 +1,12 @@
-import "./CLButtons.css"
+import "./CLButtons.css";
 
-const DEFAULT_HEIGHT = "54px"
-const DEFAULT_WIDTH = "fit-content"
-const DEFAULT_ON_CLICK = () => {}
+const DEFAULT_PRIMARY_CLASSNAME = "primary-variant-button";
+const DEFAULT_SECONDARY_CLASSNAME = "secondary-variant-button";
+
+const DEFAULT_HEIGHT = "54px";
+const DEFAULT_WIDTH = "fit-content";
+const DEFAULT_TYPE = "button";
+const DEFAULT_ON_CLICK = () => {};
 
 /**
  * PRIMARY variant of the clinical logging button.
@@ -11,24 +15,30 @@ const DEFAULT_ON_CLICK = () => {}
  */
 export const CLButtonPrimary = ({
   children,
+  className,
   height = DEFAULT_HEIGHT,
   width = DEFAULT_WIDTH,
-  onClick = DEFAULT_ON_CLICK
+  type = DEFAULT_TYPE,
+  onClick = DEFAULT_ON_CLICK,
 }) => {
+  const updatedClassName = className
+    ? DEFAULT_PRIMARY_CLASSNAME + " " + className
+    : DEFAULT_PRIMARY_CLASSNAME;
 
   return (
-    <button 
-      className="primary-variant-button"
+    <button
+      className={updatedClassName}
+      type={type}
       onClick={onClick}
       style={{
         height: height,
-        width: width
+        width: width,
       }}
     >
       {children}
     </button>
-  )
-}
+  );
+};
 
 /**
  * SECONDARY variant of the clinical logging button.
@@ -37,21 +47,27 @@ export const CLButtonPrimary = ({
  */
 export const CLButtonSecondary = ({
   children,
+  className,
   height = DEFAULT_HEIGHT,
   width = DEFAULT_WIDTH,
-  onClick = DEFAULT_ON_CLICK
+  type = DEFAULT_TYPE,
+  onClick = DEFAULT_ON_CLICK,
 }) => {
-  
+  const updatedClassName = className
+    ? DEFAULT_SECONDARY_CLASSNAME + " " + className
+    : DEFAULT_SECONDARY_CLASSNAME;
+
   return (
-    <button 
-      className="secondary-variant-button"
+    <button
+      className={updatedClassName}
+      type={type}
       onClick={onClick}
       style={{
         height: height,
-        width: width
+        width: width,
       }}
     >
       {children}
     </button>
-  )
-}
+  );
+};
