+++ frontend/src/components/NewLogModal/NewLogModal.css
@@ -1,0 +1,81 @@
+.modal-content {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 2rem;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    height: 380px;
+    width: 640px;
+    background-color: #FBFCFE;
+    border-radius: 20px;
+    box-shadow: 24;
+  }
+  
+  .close-modal-button {
+    position: absolute;
+    top: 40px;
+    left: 40px;
+    height: 40px;
+    width: 40px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: transparent;
+    color: #244B94;
+    border-radius: 40px;
+  }
+  
+  .close-modal-button:hover {
+    background-color: rgb(219, 219, 219);
+  }
+  
+  .modal-description {
+    font-size: 24px;
+    font-weight: bold;
+    margin: 1rem;
+  }
+  
+  .new-log-modal-divider {
+    width: 60%;
+  }
+  
+  .new-log-modal-buttons-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 2rem;
+    gap: 20px;
+  }
+  
+  .upload-photo-button {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    color: #F7FAFF;
+    gap: 12px;
+  }
+  
+.create-manually-button {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+  align-items: center;
+  color: #4f607e;
+  gap: 12px;
+}
+
+.close-x-icon {
+  height: 34px;
+  width: 34px;
+  color: inherit;
+}
+
+.modal-icon {
+  height: 40px;
+  width: 40px;
+}
