+++ frontend/src/pages/log_history/LogHistory.css
@@ -2,7 +2,7 @@
 .table-container {
   background: transparent;
   border-radius: 8px;
-  margin-top: 9%;
+  margin-top: 8%;
   margin-left: 6%;
   width: 90%;
 }
