+++ frontend/src/pages/upload_photo/UploadPhoto.jsx
@@ -1,15 +1,11 @@
 import { useState, useEffect } from "react";
 import { NavContentWrapper } from "../../components/NavContentWrapper/NavContentWrapper";
-import {
-  PhotoIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  ChevronDoubleRightIcon,
-  PencilSquareIcon,
-} from "@heroicons/react/24/solid";
 import ContentHeader from "../../components/ContentHeader/ContentHeader";
 import "./UploadPhoto.css";
 import { useNavigate } from "react-router-dom";
+import PreviewSection from "../../components/UploadPhoto/PreviewSection";
+import UploadArea from "../../components/UploadPhoto/UploadArea";
+import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/24/solid";
 
 export default function UploadPhoto() {
   const navigate = useNavigate();
@@ -33,37 +29,11 @@ export default function UploadPhoto() {
 }
 
 function MainContent({ handleTranscribe }) {
-  const [dragActive, setDragActive] = useState(false);
   const [files, setFiles] = useState([]);
   const [showPreview, setShowPreview] = useState(false);
-  const [isEditing, setIsEditing] = useState(false);
 
   /** Allowed file types */
   const ALLOWED_FILE_TYPES = ["image/png", "image/jpeg"];
-
-  /** Function to filter valid files */
-  const filterValidFiles = (fileList) =>
-    Array.from(fileList).filter((file) =>
-      ALLOWED_FILE_TYPES.includes(file.type)
-    );
-
-  /** Handle drag events */
-  const handleDrag = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    switch (e.type) {
-      case "dragenter":
-      case "dragover":
-        setDragActive(true);
-        break;
-      case "dragleave":
-        setDragActive(false);
-        break;
-      default:
-        break;
-    }
-  };
 
   /** Handle files from input or drop */
   const handleFiles = (newFiles) => {
@@ -75,45 +45,9 @@ function MainContent({ handleTranscribe }) {
     setFiles((prev) => [...filesWithPreview, ...prev]); // Add new files to beginning
   };
 
-  /** Handle drop event */
-  const handleDrop = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setDragActive(false);
-
-    const validFiles = filterValidFiles(e.dataTransfer.files);
-    handleFiles(validFiles);
-  };
-
-  /** Handle file input change */
-  const handleFileInput = (e) => {
-    const validFiles = filterValidFiles(e.target.files);
-    handleFiles(validFiles);
-  };
-
   /** Toggle preview visibility */
   const handlePreviewClick = () => {
     setShowPreview((prev) => !prev);
-  };
-
-  /** Toggle edit mode */
-  const handleEdit = () => {
-    setIsEditing((prev) => !prev);
-  };
-
-  /** Handle image deletion */
-  const handleDeleteImage = (timestampToDelete) => {
-    setFiles((prevFiles) => {
-      // Find the file to delete and revoke its URL
-      const fileToDelete = prevFiles.find(
-        (file) => file.timestamp === timestampToDelete
-      );
-      if (fileToDelete?.preview) {
-        URL.revokeObjectURL(fileToDelete.preview);
-      }
-      // Filter out the deleted file
-      return prevFiles.filter((file) => file.timestamp !== timestampToDelete);
-    });
   };
 
   /** Clean up object URLs when component unmounts or files change */
@@ -132,74 +66,18 @@ function MainContent({ handleTranscribe }) {
       <ContentHeader onPreviewClick={handlePreviewClick} />
 
       {showPreview && (
-        <div className="preview-section">
-          <div className="preview-header">
-            <button className="preview-title" onClick={handlePreviewClick}>
-              <ChevronDoubleRightIcon className="preview-section-icon" />
-              <span>Preview</span>
-            </button>
-            <button
-              className={`edit-button ${isEditing ? "active" : ""}`}
-              onClick={handleEdit}
-            >
-              <PencilSquareIcon className="edit-icon" />
-              <span>Edit</span>
-            </button>
-          </div>
-
-          <div className="preview-list">
-            {files.map((fileData, index) => (
-              <div key={fileData.timestamp} className="preview-item">
-                <img
-                  src={fileData.preview}
-                  alt={`Preview ${index + 1}`}
-                  className="preview-image"
-                />
-                {isEditing && (
-                  <button
-                    className="delete-button"
-                    onClick={() => handleDeleteImage(fileData.timestamp)}
-                  >
-                    ×
-                  </button>
-                )}
-              </div>
-            ))}
-          </div>
-
-          <div className="preview-footer">
-            <button className="transcribe-button" onClick={handleTranscribe}>
-              Transcribe
-            </button>
-          </div>
-        </div>
+        <PreviewSection
+          files={files}
+          setFiles={setFiles}
+          handlePreviewClick={handlePreviewClick}
+          handleTranscribe={handleTranscribe}
+        />
       )}
 
-      <div
-        className={`upload-area ${dragActive ? "drag-active" : ""}`}
-        onDragEnter={handleDrag}
-        onDragLeave={handleDrag}
-        onDragOver={handleDrag}
-        onDrop={handleDrop}
-      >
-        <input
-          type="file"
-          className="file-input"
-          accept={ALLOWED_FILE_TYPES.join(",")}
-          multiple
-          onChange={handleFileInput}
-        />
-        <label className="upload-label">
-          <div className="upload-icon">
-            <PhotoIcon className="icon-image" />
-            <div className="plus-indicator">+</div>
-          </div>
-          <div className="upload-text">
-            <h3>Upload Photos</h3>
-            <p>or drag and drop</p>
-          </div>
-        </label>
-      </div>
+      <UploadArea
+        ALLOWED_FILE_TYPES={ALLOWED_FILE_TYPES}
+        handleFiles={handleFiles}
+      />
 
       <div className="upload-controls">
         <button className="control-button prev">
